create a seeker flow for feature profile creation without backend yet
