+++ app/(onboarding)/onboarding/[role]/completed/page.tsx
@@ -1,0 +1,7 @@
+import OnboardingProfileCompleted from "@/components/OnboardingPage/OnboardingRolePage/OnboardingProfileCompleted";
+
+const page = () => {
+  return <OnboardingProfileCompleted />;
+};
+
+export default page;
