+++ components/OnboardingPage/OnboardingRolePage/ListerRolePage.tsx
@@ -1,0 +1,5 @@
+const ListerRolePage = () => {
+  return <div>ListerRolePage</div>;
+};
+
+export default ListerRolePage;
