+++ components/RoomeyForm/RommeyField.tsx
@@ -17,10 +17,10 @@ import {
 } from "@/components/ui/select";
 import { Textarea } from "@/components/ui/textarea";
 import { FieldConfig } from "@/lib/type";
-import { Check, ChevronDownIcon, CloudUpload, ImageIcon } from "lucide-react";
+import { Check, ChevronDownIcon, CloudUpload, Images } from "lucide-react";
 import Image from "next/image";
-import { useState } from "react";
-import { Control, FieldValues } from "react-hook-form";
+import { useRef, useState } from "react";
+import { Control, FieldPath, FieldValues } from "react-hook-form";
 import { Button } from "../ui/button";
 import { Calendar } from "../ui/calendar";
 import { Popover, PopoverContent, PopoverTrigger } from "../ui/popover";
@@ -33,49 +33,76 @@ type RenderFieldsProps = {
 
 const RenderFields = ({ control, fields }: RenderFieldsProps) => {
   const [openPopover, setOpenPopover] = useState<string | null>(null);
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  const pickFile = () => inputRef.current?.click();
+
+  const onContainerClick: React.MouseEventHandler<HTMLDivElement> = (e) => {
+    if (e.target === e.currentTarget) pickFile();
+  };
+
+  const onButtonPick: React.MouseEventHandler<HTMLButtonElement> = (e) => {
+    e.stopPropagation();
+    e.preventDefault();
+    pickFile();
+  };
   return fields.map((f) => {
     if (f.name === "budget_amount") {
       const unitField = fields.find((ff) => ff.name === "budget_unit");
+      const amountField = fields.find((ff) => ff.name === "budget_amount");
+
+      if (!unitField?.name || !amountField?.name) return null;
+
       return (
         <div key="budget-group" className="space-y-2">
           <FormLabel>
             {f.label} {f.required && <span>*</span>}
           </FormLabel>
+
           <div className="grid grid-cols-2 gap-3">
-            {/* amount */}
             <FormField
+              key={amountField.name}
               control={control}
-              name="budget_amount"
+              name={amountField.name as FieldPath<FieldValues>}
               render={({ field }) => (
                 <FormItem>
                   <FormControl>
                     <Input
                       type="number"
                       inputMode="decimal"
-                      min={0}
-                      placeholder={f.placeholder ?? "250"}
                       {...field}
+                      onChange={(e) =>
+                        field.onChange(
+                          e.target.value === ""
+                            ? undefined
+                            : Number(e.target.value)
+                        )
+                      }
                     />
                   </FormControl>
                   <FormMessage />
                 </FormItem>
               )}
             />
-            {/* unit */}
+
             <FormField
+              key={unitField.name}
               control={control}
-              name="budget_unit"
+              name={unitField.name as FieldPath<FieldValues>}
               render={({ field }) => (
                 <FormItem>
                   <FormControl>
-                    <Select value={field.value} onValueChange={field.onChange}>
+                    <Select
+                      onValueChange={field.onChange}
+                      defaultValue={field.value}
+                    >
                       <SelectTrigger className="w-full">
                         <SelectValue
-                          placeholder={unitField?.placeholder || "Select"}
+                          placeholder={unitField.placeholder || "Select"}
                         />
                       </SelectTrigger>
                       <SelectContent>
-                        {unitField?.options?.map((opt) => (
+                        {unitField.options?.map((opt) => (
                           <SelectItem key={opt.value} value={opt.value}>
                             {opt.label}
                           </SelectItem>
@@ -218,13 +245,12 @@ const RenderFields = ({ control, fields }: RenderFieldsProps) => {
                       value={val}
                       onChange={(e) => field.onChange(e.target.value)}
                       maxLength={max}
-                      className="h-40" // hard cap at 300
+                      className="h-40"
                       aria-describedby={`${f.name}-counter`}
                     />
                   </FormControl>
 
                   <FormDescription>
-                    {" "}
                     {`${remaining} characters remaining (min ${min}, max ${max})`}
                   </FormDescription>
 
@@ -235,7 +261,9 @@ const RenderFields = ({ control, fields }: RenderFieldsProps) => {
 
             case "radio group": {
               return (
-                <FormItem className={`${f.isRow ? "flex flex-wrap" : ""}`}>
+                <FormItem
+                  className={`${f.isRow ? "flex flex-wrap items-center" : ""}`}
+                >
                   <FormLabel className="mb-1">
                     {f.label}
                     {f.required && <span>*</span>}
@@ -275,6 +303,9 @@ const RenderFields = ({ control, fields }: RenderFieldsProps) => {
                       })}
                     </RadioGroup>
                   </FormControl>
+                  <div className="block">
+                    <FormMessage className="block bottom-0 " />
+                  </div>
                 </FormItem>
               );
             }
@@ -299,18 +330,15 @@ const RenderFields = ({ control, fields }: RenderFieldsProps) => {
                 );
               }
 
-              const inputId = `${f.name}-input`;
               const MAX_FILE_SIZE = 5 * 1024 * 1024; // 5MB
-
-              const pickFile = () => document.getElementById(inputId)?.click();
+              const inputId = `${f.name}-input`;
 
               return (
                 <FormItem>
                   <FormControl>
                     <div className="space-y-4">
-                      {/* Dropzone-style area */}
                       <div
-                        className="rounded-xl border border-dashed border-muted-foreground/30 bg-muted/20 p-6 sm:p-8 text-center"
+                        className="rounded-xl h-46 border border-dashed border-muted-foreground/30 bg-muted/20 p-6 sm:p-8 text-center"
                         onDragOver={(e) => e.preventDefault()}
                         onDrop={(e) => {
                           e.preventDefault();
@@ -322,44 +350,52 @@ const RenderFields = ({ control, fields }: RenderFieldsProps) => {
                         }}
                         role="button"
                         tabIndex={0}
-                        onClick={pickFile}
+                        onClick={onContainerClick} // <- fires only when blank area is clicked
+                        onKeyDown={(e) => {
+                          // accessibility: Enter/Space to open
+                          if (e.key === "Enter" || e.key === " ") {
+                            e.preventDefault();
+                            pickFile();
+                          }
+                        }}
                       >
-                        {!field.value ? (
+                        {!(field.value instanceof File) ? (
                           <div className="flex flex-col items-center gap-3">
-                            <ImageIcon className="h-10 w-10 text-muted-foreground" />
-                            <p className="text-sm text-muted-foreground">
-                              + Upload your best Photo
-                            </p>
+                            <Images />
+                            <p className="text-sm">+ Upload your best Photo</p>
                             <Button
                               type="button"
                               className="inline-flex items-center gap-2"
-                              onClick={pickFile}
+                              onClick={onButtonPick}
                             >
                               <CloudUpload className="h-4 w-4" />
                               Choose Photo
                             </Button>
                           </div>
                         ) : (
-                          <div className="flex flex-col items-center  gap-4">
+                          <div className="flex flex-col items-center gap-4">
                             <Image
                               src={URL.createObjectURL(field.value)}
-                              alt="Profile preview"
-                              width={500}
-                              height={500}
-                              className="h-40 w-40 rounded-full object-cover"
+                              alt="profile photo"
+                              width={100}
+                              height={100}
+                              className="rounded-full h-24 w-24"
                             />
                             <div className="flex gap-2">
                               <Button
                                 type="button"
                                 variant="outline"
-                                onClick={pickFile}
+                                onClick={onButtonPick}
                               >
                                 Replace Photo
                               </Button>
                               <Button
                                 type="button"
                                 variant="ghost"
-                                onClick={() => field.onChange(null)}
+                                onClick={(e) => {
+                                  e.stopPropagation();
+                                  field.onChange(null);
+                                }}
                               >
                                 Remove
                               </Button>
@@ -368,26 +404,30 @@ const RenderFields = ({ control, fields }: RenderFieldsProps) => {
                         )}
                         <input
                           id={inputId}
+                          name={field.name}
                           type="file"
                           accept="image/png,image/jpeg"
                           className="hidden"
+                          ref={inputRef}
                           onChange={(e) => {
                             const file = e.target.files?.[0];
+                            e.currentTarget.value = "";
+
                             if (!file) return;
                             if (!/^image\/(png|jpe?g)$/i.test(file.type))
                               return;
                             if (file.size > MAX_FILE_SIZE) return;
+
                             field.onChange(file);
                           }}
                         />
                       </div>
 
-                      {/* Tips */}
                       <div className="flex flex-col justify-center items-center">
-                        <p className="font-semibold">
+                        <p className="font-bold text-black">
                           Photo Tips for Maximum Impact:
                         </p>
-                        <ul className="mt-2 space-y-1 text-sm text-muted-foreground">
+                        <ul className="mt-2 text-sm">
                           <li className="flex items-start gap-2">
                             <Check className="mt-0.5 h-3 w-3" /> Clear, well-lit
                             face shot works best
@@ -472,7 +512,7 @@ const RenderFields = ({ control, fields }: RenderFieldsProps) => {
                           selected={selected ?? undefined}
                           captionLayout="dropdown"
                           onSelect={(d) => {
-                            field.onChange(d ?? null); // save to RHF
+                            field.onChange(d ?? null);
                             setOpenPopover(null); // close popover
                           }}
                           initialFocus
